refactor(site): reorganize demo components into engine/views/scenarios architecture
Restructure the flat demos/ directory into a three-tier architecture
that communicates the system's layering through folder structure:

- engine/ — generic playback machinery (ScenarioPlayer, Cursor, shared
  utilities) with zero Stigmer UI coupling
- views/ — Stigmer-specific UI panels (AppShell, ComposerView,
  WidgetsSidebar, SkillsListView) that render within the engine
- scenarios/ — per-scenario folders (quickstart-playback,
  skill-creation-tour, session-composer) each with index.tsx + steps.ts

Key improvements:
- Extract duplicated snapshot(), MOCK_WORKSPACE, DEMO_ORG into
  engine/shared.ts (eliminates duplication across 4 files)
- Enhance ComposerView with typingMessage/placeholder/agentRef props,
  absorbing TypingComposer so QuickstartPlayback reuses it
- Add renderWidgetsSidebar() helper to WidgetsSidebar (eliminates
  duplicated 3-line wrapper in both scenario components)
- Remove 3 dead-code files (DemoSkillCreation, skill-creation.ts,
  quickstart.ts) with zero imports
- Drop Demo prefix on internal building blocks; preserve it on
  exported MDX component names via barrel aliases

Dependencies flow one way: scenarios -> views -> engine.
Zero MDX or mdx.tsx changes required.

Made-with: Cursor
