+++ site/src/components/docs/index.ts
@@ -1,7 +1,7 @@
 export { CopyMarkdownButton } from "./copy-markdown-button";
-export { DemoQuickstartPlayback } from "./demos/DemoQuickstartPlayback";
-export { DemoSessionComposer } from "./demos/DemoSessionComposer";
-export { DemoSkillCreationTour } from "./demos/DemoSkillCreationTour";
+export { QuickstartPlayback as DemoQuickstartPlayback } from "./demos/scenarios/quickstart-playback";
+export { SessionComposerDemo as DemoSessionComposer } from "./demos/scenarios/session-composer";
+export { SkillCreationTour as DemoSkillCreationTour } from "./demos/scenarios/skill-creation-tour";
 export { glossary } from "./glossary";
 export { Mermaid } from "./mermaid";
 export { PageFeedback } from "./page-feedback";
